requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
pytest==5.3.3
requests==2.22.0
Cerberus == 1.3.1
mimesis == 3.2.0
PyMySQL == 0.9.3
selene == 2.0.0b2
openpyxl == 3.0.9